Se aplicaron refactorizaciones al main hub
