+++ app/src/main/java/com/g5/tdp2/myhealthapp/ui/HubActivity.java
@@ -2,7 +2,13 @@ package com.g5.tdp2.myhealthapp.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Context;
 import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -45,11 +51,49 @@ public class HubActivity extends AppCompatActivity {
         title.setText(getString(R.string.main_hub_title, member.getFirstname()));
 
         ListView listView = findViewById(R.id.main_hub_list_view);
-        MobileArrayAdapter listAdapter = new MobileArrayAdapter(this, VIEWS);
+        ArrayAdapter<String> listAdapter = new HubArrayAdapter(this, VIEWS);
         listView.setAdapter(listAdapter);
         listView.setOnItemClickListener((parent, view, position, id) -> {
             String selectedValue = listAdapter.getItem(position);
             Toast.makeText(HubActivity.this, selectedValue, Toast.LENGTH_SHORT).show();
         });
+    }
+}
+
+class HubArrayAdapter extends ArrayAdapter<String> {
+    private final Context context;
+    private final String[] values;
+
+    public HubArrayAdapter(Context context, String[] values) {
+        super(context, R.layout.hub_list_layout, values);
+        this.context = context;
+        this.values = values;
+    }
+
+    @Override
+    public View getView(int position, View convertView, ViewGroup parent) {
+        LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+
+        View rowView = inflater.inflate(R.layout.hub_list_layout, parent, false);
+        TextView textView = (TextView) rowView.findViewById(R.id.hub_list_label);
+        ImageView imageView = (ImageView) rowView.findViewById(R.id.hub_list_logo);
+        textView.setText(values[position]);
+
+        // Change icon based on name
+        String s = values[position];
+
+        System.out.println(s);
+
+//        if (s.equals("WindowsMobile")) {
+//            imageView.setImageResource(R.drawable.windowsmobile_logo);
+//        } else if (s.equals("iOS")) {
+//            imageView.setImageResource(R.drawable.ios_logo);
+//        } else if (s.equals("Blackberry")) {
+//            imageView.setImageResource(R.drawable.blackberry_logo);
+//        } else {
+//            imageView.setImageResource(R.drawable.android_logo);
+//        }
+
+        return rowView;
     }
 }
